+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release Notes for Blitz
 
 ## 3.7.2 - Unreleased
+### Changed
+- The values in the `excludedQueryStringParams` are now compared to query string params using regular expressions ([#272](https://github.com/putyourlightson/craft-blitz/issues/272)).
+
 ### Fixed
 - Fixed a bug that was throwing an error message if the resource base URL was set to the site URL ([#287](https://github.com/putyourlightson/craft-blitz/issues/287)).
 - Fixed a bug in which the `queue` flag was not being respected in console commands ([#294](https://github.com/putyourlightson/craft-blitz/issues/294)).